Move Wordpress and YouTube interfaces to their own folder
